bzl: Re-add library target for cargo compat
