+++ docker-images/syntax-highlighter/crates/scip-treesitter-cli/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod evaluate;
+pub mod index;
+pub mod io;
+pub mod progress;
